Cast inputs/targs to long

+++ torchtitan/datasets/fms_datasets.py
@@ -1144,7 +1144,7 @@ def build_fms_data_loader(cfg, rank, world_size):
         Perform causal language modeling by right-shifting the input sequence.
         Sets first prompt_len tokens to be ignored by the loss.
         """
-        data_seq = torch.IntTensor(data_seq)
+        data_seq = torch.LongTensor(data_seq)
         t = data_seq.clone()[1:]
         data_seq = data_seq[:-1]
         t[:prompt_len] = -100
